+++ stories/index.jsx
@@ -6,8 +6,9 @@ import {
   withKnobs,
   text,
   boolean,
-  select,
   number,
+  object,
+  select,
 } from '@storybook/addon-knobs';
 
 import DateRange from '../src/index';
@@ -20,10 +21,10 @@ storiesOf('Date range', (module))
   .add('default', () => {
     const width = text('Width', '200px');
     const translations = {
-      startDate: text('Start date', 'First day'),
-      endDate: text('End date', 'Last day'),
-      absolute: text('Absolute', 'Absolute'),
-      relative: text('Relative', 'Relative'),
+      startDate: text('Start date label', 'First day'),
+      endDate: text('End date label', 'Last day'),
+      absolute: text('Absolute label', 'Absolute'),
+      relative: text('Relative label', 'Relative'),
     };
     const relativeEnabled = boolean('Relative range enabled', false);
     const selectedRangeType = select('Range type', ['absolute', 'relative'], 'absolute');
@@ -31,6 +32,8 @@ storiesOf('Date range', (module))
     const numberOfMonths = number('Number of months', 2);
     const region = text('Region', 'en_GB');
     const showWeekNumbers = boolean('Show week numbers', true);
+    const absoluteStartDate = text('Absolute start date', '2013-04-22');
+    const absoluteEndDate = text('Absolute end date', '2013-04-22');
     const popoverProps = {
       isRelativeEnabled: relativeEnabled,
       selectedRangeType,
@@ -40,6 +43,63 @@ storiesOf('Date range', (module))
         numberOfMonths,
         region,
         showWeekNumbers,
+        endDate: absoluteEndDate,
+        startDate: absoluteStartDate,
+      },
+    };
+
+    return (
+      <DateRange
+        id="default"
+        width={width}
+        popoverProps={popoverProps}
+        onChange={action('handleChange')}
+      />
+    );
+  })
+  .add('relative range', () => {
+    const width = text('Width', '200px');
+    const translations = {
+      startDate: text('Start date label', 'First day'),
+      endDate: text('End date label', 'Last day'),
+      absolute: text('Absolute label', 'Absolute'),
+      relative: text('Relative label', 'Relative'),
+    };
+    const relativeEnabled = boolean('Relative range enabled', true);
+    const selectedRangeType = select('Range type', ['absolute', 'relative'], 'relative');
+    const dateFormat = text('Date format', 'DD.MM.YYYY');
+    const numberOfMonths = number('Number of months', 2);
+    const region = text('Region', 'en_GB');
+    const showWeekNumbers = boolean('Show week numbers', true);
+    const relativeStartDate = object(
+      'Relative start date',
+      {
+        label: 'Today',
+        value: { unit: 'DAY', timing: 'CURRENT', moment: 'START' },
+        granularity: 'DAY',
+      },
+    );
+    const relativeEndDate = object(
+      'Relative end date',
+      {
+        label: 'Tomorrow',
+        value: { unit: 'DAY', timing: 'NEXT', moment: 'END' },
+        granularity: 'DAY',
+      },
+    );
+    const popoverProps = {
+      isRelativeEnabled: relativeEnabled,
+      selectedRangeType,
+      translations,
+      absoluteRange: {
+        dateFormat,
+        numberOfMonths,
+        region,
+        showWeekNumbers,
+      },
+      relativeRange: {
+        endDate: relativeEndDate,
+        startDate: relativeStartDate,
       },
     };