+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Fixed: react-day-picker's click events don't close popover
+* Fixed: input field's default value rendering
 
 ## 1.0.0
 * Initial release